model-train-scripts/requirements.txt
# filepath: /Users/chalkiasantonios/Desktop/master-thesis/model-train-scripts/requirements.txt
# Requirements for Food Recognition and Weight Estimation model
torch>=1.13.0
torchvision>=0.14.0
pandas>=1.5.0
numpy>=1.23.0
Pillow>=9.4.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
tqdm>=4.65.0
